requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm
openpyxl
pyyaml

# Deep learning
torch>=2.0
transformers>=4.40
datasets>=2.18
accelerate>=0.28
peft>=0.10
sentence-transformers

# Existing
openai
openai-whisper
